Simplify empty fallthrough cases in switches
