+++ src/control-flow/switch-utils.ts
@@ -1,12 +1,45 @@
 import type Parser from "web-tree-sitter";
-import type { Case, EdgeType } from "./cfg-defs";
+import type { EdgeType } from "./cfg-defs";
 import type { Context } from "./generic-cfg-builder";
 import { pairwise } from "./itertools.ts";
+
 export interface SwitchOptions {
   /// A Go `select` blocks until one of the branches matches.
   /// This means that we never add an alternative edge from the
   /// head to the merge node. There is no implicit-default.
   noImplicitDefault?: boolean;
+}
+
+/**
+ * Represents a single case inside a switch statement.
+ *
+ * Cases are build of a condition and a consequence,
+ * where the condition _may_ lead to the consequence.
+ */
+export interface Case {
+  /**
+   * The condition entry node
+   */
+  conditionEntry: string;
+  /**
+   * The condition exit node
+   */
+  conditionExit: string;
+  consequenceEntry: string;
+  consequenceExit: string | null;
+  alternativeExit: string;
+  /**
+   * Does this case fall-through to the next one?
+   */
+  hasFallthrough: boolean;
+  /**
+   * Is this the default case?
+   */
+  isDefault: boolean;
+  /**
+   * A case may be entirely empty
+   */
+  isEmpty: boolean;
 }
 
 export function buildSwitch(
@@ -16,23 +49,39 @@ export function buildSwitch(
   options: SwitchOptions,
   ctx: Context,
 ) {
-  let fallthrough: string | null = null;
+  // Fallthrough from the previous case
+  let activeFallthroughs: string[] = [];
+  let hasDefaultCase = false;
   let previous: string | null = switchHeadNode;
   for (const thisCase of cases) {
     if (ctx.options.flatSwitch) {
       ctx.builder.addEdge(switchHeadNode, thisCase.conditionEntry);
-      ctx.builder.addEdge(thisCase.conditionExit, thisCase.consequenceEntry);
-      if (fallthrough) {
-        ctx.builder.addEdge(fallthrough, thisCase.consequenceEntry);
-      }
-      if (thisCase.isDefault) {
-        // If we have any default node - then we don't connect the head to the merge node.
-        previous = null;
+      if (thisCase.isEmpty && thisCase.hasFallthrough) {
+        // Since we're empty, we only care about the condition node
+        activeFallthroughs.push(thisCase.conditionExit);
+      } else {
+        ctx.builder.addEdge(thisCase.conditionExit, thisCase.consequenceEntry);
+
+        for (const fallthrough of activeFallthroughs) {
+          ctx.builder.addEdge(fallthrough, thisCase.consequenceEntry);
+        }
+        activeFallthroughs = [];
+
+        if (!thisCase.hasFallthrough && thisCase.consequenceExit) {
+          ctx.builder.addEdge(thisCase.consequenceExit, mergeNode, "regular");
+        }
+        // Update for next case
+        if (thisCase.hasFallthrough && thisCase.consequenceExit) {
+          activeFallthroughs.push(thisCase.consequenceExit);
+        }
       }
     } else {
-      if (fallthrough) {
+      /* Model the switch as an if-elif-else chain */
+      for (const fallthrough of activeFallthroughs) {
         ctx.builder.addEdge(fallthrough, thisCase.consequenceEntry);
       }
+      activeFallthroughs = [];
+
       if (previous && thisCase.conditionEntry) {
         ctx.builder.addEdge(
           previous,
@@ -50,22 +99,34 @@ export function buildSwitch(
 
       // Update for next case
       previous = thisCase.isDefault ? null : thisCase.alternativeExit;
+
+      if (!thisCase.hasFallthrough && thisCase.consequenceExit) {
+        ctx.builder.addEdge(thisCase.consequenceExit, mergeNode, "regular");
+      }
+      // Update for next case
+      if (thisCase.hasFallthrough && thisCase.consequenceExit) {
+        activeFallthroughs.push(thisCase.consequenceExit);
+      }
     }
+
+    hasDefaultCase ||= thisCase.isDefault;
 
     // Fallthrough is the same for both flat and non-flat layouts.
     if (!thisCase.hasFallthrough && thisCase.consequenceExit) {
       ctx.builder.addEdge(thisCase.consequenceExit, mergeNode, "regular");
     }
     // Update for next case
-    fallthrough = thisCase.hasFallthrough ? thisCase.consequenceExit : null;
+    if (thisCase.hasFallthrough && thisCase.consequenceExit) {
+      activeFallthroughs.push(thisCase.consequenceExit);
+    }
   }
   // Connect the last node to the merge node.
   // No need to handle `fallthrough` here as it is not allowed for the last case.
-  if (previous && !options.noImplicitDefault) {
+  if (previous && !hasDefaultCase && !options.noImplicitDefault) {
     ctx.builder.addEdge(previous, mergeNode, "alternative");
   }
 
-  if (fallthrough) {
+  for (const fallthrough of activeFallthroughs) {
     ctx.builder.addEdge(fallthrough, mergeNode, "regular");
   }
 }
@@ -113,6 +174,15 @@ export function collectCases(
       );
     }
 
+    // We want to mark empty nodes, so that we can avoid linking their
+    // consequence nodes.
+    // It is true that Go's cases are "empty" even if they have a `fallthrough`
+    // keyword as their only statement, but we can safely ignore those.
+    // That is because a Go switch allows multiple conditions, making
+    // the common case of a huge switch with many cases less common.
+    // If it comes up in practice - we'll address it.
+    const isEmpty = consequence.length === 0;
+
     cases.push({
       conditionEntry: conditionNode,
       conditionExit: conditionNode,
@@ -121,6 +191,7 @@ export function collectCases(
       alternativeExit: conditionNode,
       hasFallthrough,
       isDefault,
+      isEmpty,
     });
   }
