+++ src/test/commentTestSamples/issue77-switch-empty-fallthrough-case.go
@@ -1,0 +1,11 @@
+/* exits: 1 */
+func EmptyFallthroughCases() {
+  switch x {
+    case 1:
+        fallthrough
+    case 2:
+        fallthrough
+    case 3:
+      return;
+  }
+}
