+++ src/test/commentTestSamples/issue77-switch-empty-fallthrough-case.ts
@@ -1,0 +1,9 @@
+/* exits: 1 */
+function EmptyFallthroughCases() {
+  switch (x) {
+    case 1:
+    case 2:
+    case 3:
+      return;
+  }
+}
